# docs/swagger.yaml
basePath: /api
definitions:
...
  model.Chapter:
    properties:
      children:
        items:
          $ref: '#/definitions/model.Chapter'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      isLeaf:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      questions:
        items:
          $ref: '#/definitions/model.Question'
        type: array
      syllabus:
        $ref: '#/definitions/model.Syllabus'
      syllabusId:
        type: integer
      updatedAt:
        type: string
    type: object
...
  model.ChoiceQuestionOption:
    properties:
      content:
        type: string
      prefix:
        type: string
    type: object
...
  model.GapFillingQuestion:
    properties:
      answer:
        items:
          type: string
        type: array
    type: object
...
  model.KnowledgePoint:
    properties:
      chapter:
        $ref: '#/definitions/model.Chapter'
      chapterId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      difficulty:
        description: basic/medium/hard
        type: string
      estimatedMinutes:
        type: integer
      id:
        type: integer
      name:
        type: string
      orderIndex:
        type: integer
      questions:
        items:
          $ref: '#/definitions/model.Question'
        type: array
      updatedAt:
        type: string
    type: object
  model.KnowledgePointQuery:
    properties:
      chapterId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        type: string
      difficulty:
        type: string
      id:
        type: integer
      name:
        type: string
      pageIndex:
        type: integer
      pageSize:
        type: integer
      syllabusId:
        type: integer
      updatedAt:
        type: string
    type: object
...
  model.MultipleChoiceQuestion:
    properties:
      answer:
        items:
          type: string
        type: array
      options:
        items:
          $ref: '#/definitions/model.ChoiceQuestionOption'
        type: array
    type: object
  model.Organisation:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.PaperCode:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      name:
        type: string
      syllabus:
        $ref: '#/definitions/model.Syllabus'
      syllabusId:
        type: integer
      updatedAt:
        type: string
    type: object
...
  model.PaperSeries:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      name:
        type: string
      syllabus:
        $ref: '#/definitions/model.Syllabus'
      syllabusId:
        type: integer
      updatedAt:
        type: string
    type: object
...
  model.PastPaper:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      name:
        type: string
      paperCode:
        $ref: '#/definitions/model.PaperCode'
      paperCodeId:
        type: integer
      paperSeries:
        $ref: '#/definitions/model.PaperSeries'
      paperSeriesId:
        type: integer
      questionNumber:
        type: integer
      questions:
        items:
          $ref: '#/definitions/model.Question'
        type: array
      syllabus:
        $ref: '#/definitions/model.Syllabus'
      syllabusId:
        type: integer
      updatedAt:
        type: string
      year:
        type: integer
    type: object
...
  model.Qualification:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      name:
        type: string
      organisation:
        $ref: '#/definitions/model.Organisation'
      organisationId:
        type: integer
      updatedAt:
        type: string
    type: object
...
  model.Question:
    properties:
      chapters:
        items:
          $ref: '#/definitions/model.Chapter'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      difficult:
        type: integer
      id:
        type: integer
      indexInPastPaper:
        type: integer
      knowledgePoints:
        items:
          $ref: '#/definitions/model.KnowledgePoint'
        type: array
      pastPaper:
        $ref: '#/definitions/model.PastPaper'
      pastPaperId:
        type: integer
      questionContents:
        items:
          $ref: '#/definitions/model.QuestionContent'
        type: array
      status:
        type: integer
      stem:
        description: 原问题，接受后存入QuestionString后储存
        type: string
      syllabus:
        $ref: '#/definitions/model.Syllabus'
      syllabusId:
        type: integer
      totalScore:
        type: integer
      updatedAt:
        type: string
    type: object
...
  model.QuestionContent:
    properties:
      analyze:
        type: string
      answer:
        type: string
      gapFilling:
        $ref: '#/definitions/model.GapFillingQuestion'
      multipleChoice:
        $ref: '#/definitions/model.MultipleChoiceQuestion'
      partLabel:
        description: 部分标签
        type: string
      questionTypeId:
        type: integer
      questionTypeName:
        type: string
      score:
        type: integer
      shortAnswer:
        $ref: '#/definitions/model.ShortAnswerQuestion'
      singleChoice:
        $ref: '#/definitions/model.SingleChoiceQuestion'
      subpartLabel:
        type: string
      trueOrFalse:
        $ref: '#/definitions/model.TrueOrFalseQuestion'
    type: object
...
  model.ShortAnswerQuestion:
    properties:
      answer:
        type: string
    type: object
  model.SingleChoiceQuestion:
    properties:
      answer:
        type: string
      options:
        items:
          $ref: '#/definitions/model.ChoiceQuestionOption'
        type: array
    type: object
...
  model.Syllabus:
    properties:
      code:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      name:
        type: string
      qualification:
        $ref: '#/definitions/model.Qualification'
      qualificationId:
        type: integer
      updatedAt:
        type: string
    type: object
...
  model.TrueOrFalseQuestion:
    properties:
      answer:
        description: 1 正确 2 错误
        type: integer
    type: object
...
info:
  contact:
    name: LTEdu Support
  description: LTEdu 在线教育平台后端接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: LTEdu API
  version: "1.0"
paths:
...
  /v1/chapter/generate-keypoints:
    post:
      consumes:
      - application/json
      description: 为指定章节自动生成知识点
      parameters:
      - description: 章节ID和模式
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 生成章节知识点
      tags:
      - 知识点
...
  /v1/knowledge-point/byChapter:
    post:
      consumes:
      - application/json
      description: 根据章节ID获取该章节下的所有知识点
      parameters:
      - description: 章节ID
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 根据章节获取知识点
      tags:
      - 知识点
  /v1/knowledge-point/byId:
    post:
      consumes:
      - application/json
      description: 根据知识点ID获取知识点详情
      parameters:
      - description: 知识点ID
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 根据ID获取知识点
      tags:
      - 知识点
  /v1/knowledge-point/bySyllabus:
    post:
      consumes:
      - application/json
      description: 根据考纲ID获取该考纲下的所有知识点
      parameters:
      - description: 考纲ID
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 根据考纲获取知识点
      tags:
      - 知识点
  /v1/knowledge-point/create:
    post:
      consumes:
      - application/json
      description: 创建新知识点
      parameters:
      - description: 知识点信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.KnowledgePoint'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建知识点
      tags:
      - 知识点
  /v1/knowledge-point/delete:
    post:
      consumes:
      - application/json
      description: 删除指定知识点
      parameters:
      - description: 知识点ID
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除知识点
      tags:
      - 知识点
  /v1/knowledge-point/edit:
    post:
      consumes:
      - application/json
      description: 更新现有知识点信息
      parameters:
      - description: 知识点信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.KnowledgePoint'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新知识点
      tags:
      - 知识点
  /v1/knowledge-point/list:
    post:
      consumes:
      - application/json
      description: 分页查询知识点列表
      parameters:
      - description: 查询条件
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.KnowledgePointQuery'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取知识点列表
      tags:
      - 知识点
...
  /v1/question/auto-link-keypoints:
    post:
      consumes:
      - application/json
      description: 将题目自动关联到匹配的知识点
      parameters:
      - description: 题目ID、章节ID和考纲ID
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 自动关联题目到知识点
      tags:
      - 知识点
  /v1/question/auto-link-keypoints-intelligent:
    post:
      consumes:
      - application/json
      description: 使用两阶段智能方法将题目关联到知识点
      parameters:
      - description: 题目ID和考纲ID
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 智能关联题目到知识点
      tags:
      - 知识点
...
  /v1/syllabus/auto-migrate-keypoints:
    post:
      consumes:
      - application/json
      description: 批量自动化迁移考纲知识点（仅管理员）
      parameters:
      - description: 考纲ID和迁移选项
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 批量自动迁移考纲
      tags:
      - 知识点
...
securityDefinitions:
  BearerAuth:
    description: '输入 "Bearer {token}" 格式的 JWT，例如: Bearer eyJhbGci...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"